vgt: use 64-bit pa in vgt_emulate_read/write()
When a vGT VM enables PPGTT, its PPGTT pagetable page's address can be
over 4GB, so vgt_emulate_write() needs a 64-bit pa.

BTW, in the current implementation of vPPGTT, HVM VM's reading
PPGTT pagetable page is not trapped to vGT at all, so actually the pa
parameter of vgt_emulate_read() can be uint32_t.
However, let's change it to uint64 too for consistency with
vgt_emulate_write().

+++ drivers/xen/vgt/mmio.c
@@ -258,7 +258,8 @@ bool default_passthrough_mmio_read(struct vgt_device *vgt, unsigned int offset,
 
 #define PCI_BAR_ADDR_MASK (~0xFUL)  /* 4 LSB bits are not address */
 
-static inline unsigned int vgt_pa_to_mmio_offset(struct vgt_device *vgt, unsigned long pa)
+static inline unsigned int vgt_pa_to_mmio_offset(struct vgt_device *vgt,
+	uint64_t pa)
 {
 	return (vgt->vm_id == 0)?
 		pa - vgt->pdev->gttmmio_base :
@@ -270,7 +271,7 @@ static inline unsigned int vgt_pa_to_mmio_offset(struct vgt_device *vgt, unsigne
  * Emulate the VGT MMIO register read ops.
  * Return : true/false
  * */
-bool vgt_emulate_read(struct vgt_device *vgt, unsigned int pa, void *p_data,int bytes)
+bool vgt_emulate_read(struct vgt_device *vgt, uint64_t pa, void *p_data,int bytes)
 {
 	struct vgt_mmio_entry *mht;
 	struct pgt_device *pdev = vgt->pdev;
@@ -327,7 +328,7 @@ bool vgt_emulate_read(struct vgt_device *vgt, unsigned int pa, void *p_data,int 
  * Emulate the VGT MMIO register write ops.
  * Return : true/false
  * */
-bool vgt_emulate_write(struct vgt_device *vgt, unsigned int pa,
+bool vgt_emulate_write(struct vgt_device *vgt, uint64_t pa,
 	void *p_data, int bytes)
 {
 	struct pgt_device *pdev = vgt->pdev;
